+++ CHANGELOG.md
@@ -1,5 +1,18 @@
 # Released Notes
 
+## v1.1.0 - (2020-07-27)
+
+### Added
+
+- Added `customPagination` class
+- Added new checks to the `insert`,` update` and `delete` methods
+
+### Fixed
+
+- Fixed bugs in the `CRUD` class
+- Fixed connection to the database
+- Fixed PSR-19
+------------------------------------------------------------------------
 ## v1.0.0 - (2020-07-17)
 
 ### Changed